Tests for task 8 New properties using getters and setters to set and check values.
