+++ student/src/main/java/com/bazzaware/Student.java
@@ -6,16 +6,18 @@ public class Student {
 
     private String _lastName;
     private String _firstName;
+    private String _course;
     private Double _mark;
     private String _id;
 
-    public Student(String FirstName, String LastName, Double Mark) {
-        _firstName = FirstName;
-        _lastName = LastName;
-        if (Mark < 1 || Mark > 100) {
+    public Student(String firstName, String lastName, String course, Double mark) {
+        _firstName = firstName;
+        _lastName = lastName;
+        _course = course;
+        if (mark < 1 || mark > 100) {
             throw new IllegalArgumentException("Mark must be between 1 and 100");
         }
-        _mark = Mark;
+        _mark = mark;
         _id = UUID.randomUUID().toString();
     }
 
@@ -27,7 +29,7 @@ public class Student {
     public String ToString() {
         String name = _firstName + " " + _lastName;
         String mark = _mark.toString();
-        String result = "Name: " + name + " Mark: " + mark;
+        String result = "Name: " + name + "\nCourse: " + _course + "\nMark: " + mark;
         return result;
     }
